microservices/InterviewProcessing/requirements.txt
json
google.generativeai
pymongo
uuid
pika